pyproject.toml
[tool.poetry]
name = "drf-project"
version = "0.1.0"
description = ""
authors = ["Kim Gwang woon <misili304@naver.com>"]
readme = "README.md"
packages = [{include = "drf_project"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2"
django-cors-headers = "^3.14.0"
djangorestframework = "^3.14.0"
djangorestframework-simplejwt = "^5.2.2"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
